fixing dependencys and client structure
