+++ client/src/main/scala/uk/gov/ons/addressIndex/client/AddressIndexClient.scala
@@ -1,39 +1,36 @@
 package uk.gov.ons.addressIndex.client
 
 import scala.concurrent.Future
-import uk.gov.ons.addressIndex.client.AddressIndexClient.{AddressIndexServerHost, UprnQuery}
-import uk.gov.ons.addressIndex.model.{AddressIndexSearchRequest, AddressIndexUPRNRequest, AddressScheme}
-import play.api.libs.ws.{WSClient, WSRequest}
+import uk.gov.ons.addressIndex.model.{AddressIndexSearchRequest, AddressIndexUPRNRequest}
+import play.api.libs.ws.{WSClient, WSRequest, WSResponse}
+import uk.gov.ons.addressIndex.client.AddressIndexClientHelper.{AddressIndexServerHost, AddressQuery, UprnQuery}
 
 trait AddressIndexClient {
 
   /**
     * @return a standard web service client
     */
-  def client: WSClient
+  def client : WSClient
 
   /**
     * @return the host address of the address index server
     */
-  def host: String
-  
+  def host : String
+
   protected implicit lazy val iClient : WSClient = client
-  protected implicit lazy val iHost : AddressIndexServerHost = AddressIndexServerHost(host)
+  protected implicit lazy val iHost : AddressIndexServerHost = host
 
   /**
     * perform an address search query
     *
-    * @param req the request
+    * @param request the request
     * @return a list of addresses
     */
-  def addressQuery(req : AddressIndexSearchRequest) : Future[_] = {
+  def addressQuery(request : AddressIndexSearchRequest) : Future[WSResponse] = {
     AddressQuery
-      .toReq
       .withQueryString(
-        Seq(
-          "input" -> req.input,
-          "format" -> req.format.toString
-        )
+        "input" -> request.input,
+        "format" -> request.format.toString
       )
       .get
   }
@@ -41,36 +38,45 @@ trait AddressIndexClient {
   /**
     * perform a uprn query
     *
-    * @param req the request
+    * @param request the request
     * @return an address
     */
-  def uprnQuery(req : AddressIndexUPRNRequest) : Future[_] = {
-    UprnQuery(req.uprn.toString)
-      .toReq
-      .withQueryString(Seq("format" -> req.format.toString))
+  def uprnQuery(request : AddressIndexUPRNRequest) : Future[WSResponse] = {
+    UprnQuery(request.uprn.toString)
+      .withQueryString(
+        "format" -> request.format.toString
+      )
       .get
   }
 }
 
-object AddressIndexClient {
+object AddressIndexClientHelper {
+
+  implicit def str2host(h : String) : AddressIndexServerHost = AddressIndexServerHost(h)
+
+  implicit def path2req(p : AddressIndexPath)(implicit client : WSClient, host : AddressIndexServerHost) : WSRequest =
+    client url s"${host.value}${p.path}" withMethod p.method
 
   sealed abstract class AddressIndexPath(val path : String, val method : String)
 
   case class AddressIndexServerHost(value : String)
 
-  implicit class AddressIndexPathToWsAugmenter(h : AddressIndexPath)(implicit client : WSClient) {
-    def toReq() : WSRequest = {
-      client url s"${host.value}${p.path}" withMethod p.path
-    }
-  }
-
-  object UprnQuery extends AddressIndexPath("", "") {
+  object UprnQuery extends AddressIndexPath(
+    path = "",
+    method = ""
+  ) {
     def apply(uprn : String) = {
       val initialRoute = "/addresses"
       val fullRoute = s"$initialRoute/$uprn"
-      new AddressIndexPath(fullRoute, "GET") {}
+      new AddressIndexPath(
+        path = fullRoute,
+        method = "GET"
+      ) {}
     }
   }
 
-  object AddressQuery extends AddressIndexPath("/addresses", "GET")
+  object AddressQuery extends AddressIndexPath(
+    path = "/addresses",
+    method = "GET"
+  )
 }
